requirements.txt
feedparser>=6.0.11
requests>=2.32.3
python-dotenv>=1.0.1
google-genai>=0.3.0
beautifulsoup4>=4.12.3
deep-translator>=1.11.4
tzdata>=2026.1
yfinance>=0.2.40
pytest>=8.0.0
